% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tnm_data}
\alias{tnm_data}
\title{Retrieve and Parse Data from a Thenmap Dataset}
\usage{
tnm_data(
  dataset,
  date = "*",
  data_props = c("id", "name"),
  language = tnm_get_config()$language,
  version = tnm_get_config()$version
)
}
\arguments{
\item{dataset}{Name or ID of the Thenmap dataset.}

\item{date}{Optional date (default "*").}

\item{data_props}{Vector of requested data property names (default \code{c("id", "name")}).}

\item{language}{Language code (default from config).}

\item{version}{Dataset version (default from config).}
}
\value{
A  data frame or list with the downloaded dataset.
}
\description{
Retrieve and Parse Data from a Thenmap Dataset
}
