% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{tnm_geo}
\alias{tnm_geo}
\title{Retrieve Geospatial Data from a Thenmap Dataset}
\usage{
tnm_geo(
  dataset,
  date,
  geo_type = c("geojson", "topojson"),
  geo_crs = "wgs84",
  geo_props = NULL,
  language = tnm_get_config()$language,
  version = tnm_get_config()$version
)
}
\arguments{
\item{dataset}{Name or ID of the Thenmap dataset.}

\item{date}{Date to retrieve data for.}

\item{geo_type}{Geospatial format: \code{"geojson"} or \code{"topojson"} (default \code{"geojson"}).}

\item{geo_crs}{Coordinate reference system, e.g., \code{"wgs84"} (default).}

\item{geo_props}{Vector of requested geographic property names (optional).}

\item{language}{Language code (default from config).}

\item{version}{Dataset version (default from config).}
}
\value{
An \code{{sf}} object or parsed geospatial list.
}
\description{
Retrieve Geospatial Data from a Thenmap Dataset
}
