% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{tnm_info}
\alias{tnm_info}
\title{Retrieve Metadata Information About a TNM Dataset}
\usage{
tnm_info(
  dataset,
  version = tnm_get_config()$version,
  language = tnm_get_config()$language
)
}
\arguments{
\item{dataset}{Name or ID of the TNM dataset.}

\item{version}{Dataset version (default from config).}

\item{language}{Language code (default from config).}
}
\value{
A list containing metadata and properties of the dataset.
}
\description{
Retrieve Metadata Information About a TNM Dataset
}
