% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{tnm_set_config}
\alias{tnm_set_config}
\title{Set a custom configuration for the thenmap API client}
\usage{
tnm_set_config(
  base_url = NULL,
  version = NULL,
  language = NULL,
  timeout = NULL
)
}
\arguments{
\item{base_url}{URL used for the API endpoint}

\item{version}{Version of the API}

\item{language}{Default language to be used for API calls}

\item{timeout}{Default timeout to exit http request}
}
\description{
Set a custom configuration for the thenmap API client
}
