% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{tnm_svg}
\alias{tnm_svg}
\title{Retrieve an SVG Map from a TNM Dataset}
\usage{
tnm_svg(
  dataset,
  date,
  svg_proj = NULL,
  svg_width = 600,
  svg_height = 600,
  svg_props = NULL,
  language = tnm_get_config()$language,
  version = tnm_get_config()$version
)
}
\arguments{
\item{dataset}{Name or ID of the TNM dataset.}

\item{date}{Date to retrieve data for.}

\item{svg_proj}{Optional projection name for the SVG.}

\item{svg_width}{Width of the SVG in pixels (default 600).}

\item{svg_height}{Height of the SVG in pixels (default 600).}

\item{svg_props}{Optional vector of SVG property names to include.}

\item{language}{Language code (default from config).}

\item{version}{Dataset version (default from config).}
}
\value{
SVG string
}
\description{
Retrieve an SVG Map from a TNM Dataset
}
